move assets related stuff to "app/assets" folder
