--- app/__init__.py
@@ -1,5 +1,0 @@
-from .api.assets_routes import register_assets_system
-from .assets_scanner import sync_seed_assets
-from .database.db import init_db_engine
-
-__all__ = ["init_db_engine", "sync_seed_assets", "register_assets_system"]
