+++ app/assets/database/helpers/bulk_ops.py
