+++ app/assets/database/helpers/escape_like.py
