+++ app/assets/database/helpers/fast_check.py
