+++ app/assets/database/helpers/ownership.py
