+++ app/assets/manager.py
@@ -6,13 +6,13 @@ from typing import Optional, Sequence
 
 from comfy_api.internal import async_to_sync
 
-from ._assets_helpers import (
+from ..db import create_session
+from ._helpers import (
     ensure_within_base,
     get_name_and_tags_from_asset_path,
     resolve_destination_from_tags,
 )
 from .api import schemas_in, schemas_out
-from .database.db import create_session
 from .database.models import Asset
 from .database.services import (
     add_tags_to_asset_info,
